+++ udata_front/frontend/helpers.py
@@ -446,15 +446,3 @@ def visibles(value):
     if not isinstance(value, list):
         raise ValueError('visibles only accept list as parameter')
     return list(filter(lambda elt: elt.is_visible, value))
-
-
-@front.app_template_filter()
-@contextfilter
-def permissions(ctx, resources):
-    '''Return permissions for resources'''
-    permissions = {}
-    for resource in resources:
-        element_to_check = resource if resource.from_community else resource.dataset
-        can_edit_resource = ctx['can_edit_resource']
-        permissions[str(resource.id)] = can_edit_resource(element_to_check).can()
-    return permissions
